Put the extra input composables in separate files
Fixes #691

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/enaboapps/Switchify/issues/691?shareId=XXXX-XXXX-XXXX-XXXX).
