+++ app/src/main/java/com/enaboapps/switchify/screens/settings/actions/inputs/CopyTextExtraInput.kt
@@ -1,0 +1,28 @@
+package com.enaboapps.switchify.screens.settings.actions.inputs
+
+import androidx.compose.runtime.Composable
+import com.enaboapps.switchify.service.custom.actions.store.data.ActionExtra
+import com.enaboapps.switchify.widgets.TextArea
+
+@Composable
+fun CopyTextExtraInput(
+    selectedExtra: ActionExtra?,
+    onExtraUpdated: (ActionExtra?) -> Unit,
+    onExtraValidated: (Boolean) -> Unit
+) {
+    TextArea(
+        value = selectedExtra?.textToCopy ?: "",
+        onValueChange = { text ->
+            onExtraUpdated(
+                ActionExtra(
+                    textToCopy = text
+                )
+            )
+
+            onExtraValidated(text.isNotBlank())
+        },
+        label = "Text to Copy",
+        isError = selectedExtra?.textToCopy.isNullOrBlank() == true,
+        supportingText = "Text to copy is required"
+    )
+}
